add asset bulk ev4

+++ app/hyposoft/equipment/resources.py
@@ -6,11 +6,11 @@ from import_export.resources import ModelResource
 
 from equipment.handlers import create_itmodel_extra, create_asset_extra
 from changeplan.models import ChangePlan
-from hyposoft.utils import versioned_object, add_asset, add_rack
+from hyposoft.utils import versioned_object, add_asset, add_rack, newest_object
 from .models import ITModel, Asset, Rack, Site
 from network.models import NetworkPortLabel
 from power.models import Powered, PDU
-from import_export.widgets import ForeignKeyWidget
+from import_export.widgets import ForeignKeyWidget, Widget, NumberWidget, IntegerWidget
 
 import re
 
@@ -101,7 +101,7 @@ class ITModelResource(ModelResource):
 
     class Meta:
         model = ITModel
-        exclude = 'id'
+        exclude = 'id', 'type'
         import_id_fields = ('vendor', 'model_number')
         export_order = ('mount_type', 'vendor', 'model_number', 'height', 'display_color', 'network_ports',
                         'power_ports', 'cpu', 'memory', 'storage', 'comment',
@@ -109,6 +109,24 @@ class ITModelResource(ModelResource):
         skip_unchanged = True
         report_skipped = True
         clean_model_instances = True
+
+    def before_import_row(self, row, **kwargs):
+        if row['mount_type'] == 'asset':
+            row['mount_type'] = 'regular'
+
+        if row['mount_type'] == ITModel.Type.BLADE:
+            row['height'] = 9
+            row['power_ports'] = 0
+            row['network_ports'] = 1
+            row['network_port_name_1'] = 'chassis'
+
+    def after_export(self, queryset, data, *args, **kwargs):
+        for i, row in enumerate(data):
+            del data[i]
+            row = list(row)
+            if row[0] == "regular":
+                row[0] = "asset"
+            data.insert(i, row)
 
     def after_import_row(self, row, row_result, **kwargs):
         my_model = ITModel.objects.get(vendor=row['vendor'], model_number=row['model_number'])
@@ -128,17 +146,38 @@ class ITModelResource(ModelResource):
         create_itmodel_extra(my_model, special + current)
 
 
+def get_site(row):
+    if row:
+        if row['datacenter'] and row['rack'] and row['rack_position']:
+            return row['datacenter']
+        elif row['offline_site']:
+            return row['offline_site']
+        elif row['chassis_number']:
+            chassis = newest_object(
+                Asset, ChangePlan.objects.get(id=row['version']), asset_number=row['chassis_number']
+            )
+            return chassis.site
+    return None
+
+
 class AssetResource(VersionedResource):
     class RackForeignKeyWidget(ForeignKeyWidget):
-        def clean(self, value, row):
-            my_rack = row['rack']
-            my_datacenter = Site.objects.get(abbr=row['datacenter'])
-            rack = Rack.objects.filter(rack=my_rack, site=my_datacenter)\
-                .order_by("-version__id").first()
-            return add_rack(rack, ChangePlan.objects.get(id=row['version']))
+        def clean(self, value, row=None, *args, **kwargs):
+            if row['datacenter']:
+                my_rack = row['rack']
+                my_site = get_site(row)
+                version = ChangePlan.objects.get(id=row['version'])
+                if my_site:
+                    site = Site.objects.get(abbr=my_site)
+                    rack = newest_object(Rack, version, rack=my_rack, site=site)
+                    return add_rack(rack, version)
+            return None
+
+        def render(self, value, obj=None):
+            return '' if obj and obj.itmodel.type == ITModel.Type.BLADE else super().render(value, obj)
 
     class ITModelForeignKeyWidget(ForeignKeyWidget):
-        def clean(self, value, row):
+        def clean(self, value, row=None, *args, **kwargs):
             return self.model.objects.get(
                 vendor__iexact=row['vendor'],
                 model_number__iexact=row['model_number']
@@ -151,11 +190,32 @@ class AssetResource(VersionedResource):
             else:
                 return None
 
+    class SiteWidget(ForeignKeyWidget):
+        def __init__(self, column, *args, **kwargs):
+            self.column = column
+            super().__init__(Site, 'abbr', *args, **kwargs)
+
+        def clean(self, value, row=None, *args, **kwargs):
+            site = get_site(row)
+            return super().clean(site, row, *args, **kwargs)
+
+        def render(self, value, obj=None):
+            if value and obj and not obj.itmodel.type == ITModel.Type.BLADE:
+                if self.column == 'datacenter' and not value.offline:
+                    return value.abbr
+                if self.column == 'offline_site' and value.offline:
+                    return value.abbr
+            return ""
 
     datacenter = fields.Field(
         column_name='datacenter',
         attribute='site',
-        widget=ForeignKeyWidget(Site, 'abbr')
+        widget=SiteWidget('datacenter')
+    )
+    offline_site = fields.Field(
+        column_name='offline_site',
+        attribute='site',
+        widget=SiteWidget('offline_site')
     )
     rack = fields.Field(
         column_name='rack',
@@ -182,6 +242,33 @@ class AssetResource(VersionedResource):
     version = fields.Field(
         attribute='version',
         widget=VersionWidget(ChangePlan, 'id')
+    )
+    blade_chassis = fields.Field(
+        column_name='chassis_number',
+        attribute='blade_chassis',
+        widget=ForeignKeyWidget(Asset, 'asset_number')
+    )
+    slot = fields.Field(
+        column_name='slot_number',
+        attribute='slot',
+        widget=IntegerWidget()
+    )
+    display_color = fields.Field(
+        column_name='custom_display_color',
+        attribute='display_color',
+    )
+    cpu = fields.Field(
+        column_name='custom_cpu',
+        attribute='cpu',
+    )
+    memory = fields.Field(
+        column_name='custom_memory',
+        attribute='memory',
+        widget=IntegerWidget()
+    )
+    storage = fields.Field(
+        column_name='custom_storage',
+        attribute='storage',
     )
 
     def skip_row(self, instance, original):
@@ -231,10 +318,12 @@ class AssetResource(VersionedResource):
 
     class Meta:
         model = Asset
-        exclude = ('id', 'itmodel', 'commissioned', 'decommissioned_by', 'decommissioned_timestamp')
+        exclude = ('id', 'itmodel', 'site', 'commissioned', 'decommissioned_by', 'decommissioned_timestamp')
         import_id_fields = ('hostname', 'vendor', 'model_number')
-        export_order = ('asset_number', 'datacenter', 'hostname', 'rack', 'rack_position', 'vendor', 'model_number',
-                        'owner', 'comment', 'power_port_connection_1', 'power_port_connection_2')
+        export_order = ('asset_number', 'hostname', 'datacenter', 'offline_site', 'rack', 'rack_position',
+                        'blade_chassis', 'slot', 'vendor', 'model_number',
+                        'owner', 'comment', 'power_port_connection_1', 'power_port_connection_2',
+                        'display_color', 'cpu', 'memory', 'storage')
         skip_unchanged = True
         report_skipped = True
         clean_model_instances = True
@@ -242,7 +331,7 @@ class AssetResource(VersionedResource):
     def before_import_row(self, row, **kwargs):
         if row['asset_number'] == '' and self.version.id == 0:
             try:
-                exists = Asset.objects.get(hostname=row['hostname'])
+                exists = Asset.objects.get(hostname=row['hostname'], version_id=0)
                 row['asset_number'] = exists.asset_number
             except:
                 max_an = Asset.objects.all().aggregate(Max('asset_number'))
@@ -259,48 +348,49 @@ class AssetResource(VersionedResource):
             return  # skip
 
         my_asset = add_asset(my_asset, self.version)
-        my_rack = my_asset.rack
 
-        special = []
-        for i in range(1, min(3, my_asset.itmodel.power_ports + 1)):
-            pc = row['power_port_connection_' + str(i)]
-            if len(pc) > 0:
-                try:
-                    split = re.search(r"\d", pc).start()
-                    position = pc[:split]
-                    plug = int(pc[split:])
-                    special.append({'pdu_id': PDU.objects.get(rack=my_rack, position=position), 'plug': plug})
-                except AttributeError:
-                    raise ValidationError('power_port_connection_' + str(i) + "formatted incorrectly")
-                except PDU.DoesNotExist:
-                    raise ValidationError(pc + " does not exit on the specified rack")
-        current = []
-        if len(special) >= 2:
-            special_simple = [port['pdu_id'].position + str(port['plug']) for port in special]
+        if not my_asset.site.offline and not my_asset.itmodel.type == ITModel.Type.BLADE:
+            my_rack = my_asset.rack
+            special = []
+            for i in range(1, min(3, my_asset.itmodel.power_ports + 1)):
+                pc = row['power_port_connection_' + str(i)]
+                if len(pc) > 0:
+                    try:
+                        split = re.search(r"\d", pc).start()
+                        position = pc[:split]
+                        plug = int(pc[split:])
+                        special.append({'pdu_id': PDU.objects.get(rack=my_rack, position=position), 'plug': plug})
+                    except AttributeError:
+                        raise ValidationError('power_port_connection_' + str(i) + "formatted incorrectly")
+                    except PDU.DoesNotExist:
+                        raise ValidationError(pc + " does not exit on the specified rack")
+            current = []
+            if len(special) >= 2:
+                special_simple = [port['pdu_id'].position + str(port['plug']) for port in special]
 
-            current = [{'pdu_id': port['pdu'], 'plug': port['plug_number']}
-                       for port in my_asset.powered_set.order_by('order').values('pdu', 'plug_number')]
+                current = [{'pdu_id': port['pdu'], 'plug': port['plug_number']}
+                           for port in my_asset.powered_set.order_by('order').values('pdu', 'plug_number')]
 
-            for port in current:
-                simple = port['pdu_id'].position + str(port['plug'])
-                if simple in special_simple:
-                    current.remove(simple)
+                for port in current:
+                    simple = port['pdu_id'].position + str(port['plug'])
+                    if simple in special_simple:
+                        current.remove(simple)
 
-        my_asset.powered_set.all().delete()
+            my_asset.powered_set.all().delete()
 
-        if row_result.import_type == "new":
-            ports = [{"label": port, "mac_address": None, "connection": None}
-                     for port in
-                     NetworkPortLabel.objects.filter(itmodel=my_asset.itmodel).values_list('name', flat=True)]
-        else:
-            ports = None
-        my_asset.version = self.version
-        create_asset_extra(
-            my_asset,
-            self.version,
-            special + current,
-            ports
-        )
+            if row_result.import_type == "new":
+                ports = [{"label": port, "mac_address": None, "connection": None}
+                         for port in
+                         NetworkPortLabel.objects.filter(itmodel=my_asset.itmodel).values_list('name', flat=True)]
+            else:
+                ports = None
+            my_asset.version = self.version
+            create_asset_extra(
+                my_asset,
+                self.version,
+                special + current,
+                ports
+            )
 
     def after_export(self, queryset, data, *args, **kwargs):
         del data['version']
